Set up the beginnings of a working prototype
This sets up the authorization mechanisms that are required in order to
ask a user for permission to get bits information from their channel,
handle errors, talk to the Twitch API to fetch the information and
display it in the console.

The actual work on PubSub is currently still left as an exercise to the
reader.