Refactor the code to be more betterer
This takes the original example and tweaks it so that the files that are
displayed are transmitted via an actual server. We also have added in
two template files and set up the application to serve them statically.

This will allow for us to easily make changes without constantly having
to work with templates and stubs as we did previously.