Include initial WebSocket support using ws-wrapper
Now that we have parcel support working, we can set up the websocket
code stubs, which are currently set to establish connections and send
simple messages between the client and the server.

This is implemented currently as the client sending a message
immediately on connection, so that we can verify.