Add the ability to drop authorization
Since we have the ability for the user to associate their account with
the overlay to get it to work, they should also have the ability to
drop that authorization should they want or need to.

In that regard, we now have a /deauth endpoint that will redact any 
authorization that's currently in place.

This also fixes a bug involving multiple copies of the same event
triggering because previously we always set up a new listener without
first removing the old one.