Use our websocket connection to transmit auth
This modifies the control panel to include a button that fills the width
of the page and which changes its text and link based on the current
state of authorization on the back end.

The back end now broadcasts authorization changes as they happen and
makes sure to transmit one to any newly connected or reconnected socket
so that the information is always available.