Tweak the CSS colors for the control panel
This includes a couple of minor rules to tweak the colors used a bit.

In particular, text is white and not black so that it's visible, and
there is a background color set that mimics the appropriate background
color so that the panel appears correctly in slobs (in as much as it
looks like other panels).

The background color here mimics the colors used by the eventual user of
the panel, while the commented out color is the normal background of a
dark mode slobs.

There is probably some method by which you could obtain the css of the
theme from slobs, but if so I don't know what it is.