Centralize websocket init and include reconnection
This does two important things; it stops the code duplication that we
had in regards to websocket setup by using a "library" that will set
the appropriate event handlers for us.

Secondly, one of those event handlers will get the socket to reconnect
itself should it have a connection failure.