Include Twitch chat support
One of the things that would make the overlay extra useful would be if 
it can talk to chat. This could be used for status messages and updates,
such as reminders, messages when new people enter the fray or take a new
position on the leaderboard, etc.

It would also be handy to allow the overlay to tangentially trigger bot
commands in other bots in reaction to events, such as a command that 
could connect a channel point redeem with a bot command that does 
something interesting.