Include some chat helper functions
In order to make the chat message and action sending easier to manage,
this includes some simple helpers that will allow you to do both things
without needing to access the Twurple chat client object directly. 

Both are smart enough to not do anything if the user isn't currently
authorized or the chat is not set up.