Flesh out the testing code for more robust data
This change fleshes out our ability to test things by putting more
controls into the test form and making sure that the test data is as
accurate as possible (more to come in a future commit in this area).

In particular this means that we can throw bits anonymously or not, and
subscribe anonymously or not, as well as switch subscriptions between
gifts and regular subs.

This is the next to last step in actually tracking the leaderboard.