Clarify and clean up the test panel
This cleans up the page that's used for generating test messages to make
it more robust and easier to use.

The larger part of this change is that we have a list of users that we
can pick from instead of a text field, and that allows us to gather the
information on a user (name, display name and user id) that we need to
fake up messages in a believable way on the back end.

Additionally the messages that transmit are more "correct" than they
were before by ensuring that you can't anonymously sub (for example) and
other similar cleanups.