Reorganize the code
Now that we're getting closer to implementing some actual solutions to
the overall overlay and its data, it's time to do a wee bit of
refactoring on the code to change the layout and structure.

Until now it was some simply laid out files that were meant just to test
the pubsub library and connectivity, which organically grew into the
code that currently exists and badly needs to be rearranged.

There's still more to do in this regard; this one largely just touches 
the client code, the single file server code should be made more robust
as well.