Use a proper configuration system
This changes how we get our configuration information so that instead of
a bunch of hard coded nonsense, instead we can load data from a
configuration file on disk (optionally; does not have to be there) as
well as environment variables.

The configuration system is powered by convict, which is a fairly
powerful way to use a schema to provide configuration information,
including default values and a priority system that lets you override
configuration files with environment variables, etc.

All of the back end code has been modified to use the new configuration
system. In addition the front end code needed to be slightly tweaked so
as to be able to capture the configuration information it needs re:
ports from the back end.