Ensure that old tokens don't cause issues
It's possible that while the app is not running that you deauthorize it
from your twitch account while it has a stored token. In that case when
the app starts and tries to refresh the token, it gets kicked right in
the yams.

This fixes that by more robustly detecting when the token is invalid and
gracefully makes sure that Twitch support doesn't get set up and that
the offending token is removed from the database.

It's currently a little unclear what happens if you deauthorize the app
while it's actively running and it hits a point where it tries to
refresh the token and fails. So, cowardly, until we know how that
manifests we're going to assume nobody is going to do that while
they're actively running a giveaway.