Include new database schemes for our giveaway work
In order to run a giveaway we need to be able to track information about
them, such as their history, when they started and ended and who
participated in them.

Further, for participants we need to know which giveways they
participated in (which could be many) and what they did as their
participation in each giveaway.

To that end, this includes the database schemas for storing this
information.