Include some of the back end logic for giveaways
This sets up the beginnings of the back end code that are required in
order for the two ends to communicate wether or not there's currently a
giveaway in progress or not.

This also includes the start of the front end controls that will allow
the back end to know if it should start or stop a giveaway, etc.