Fix an issue with giveaway tracking
We had a slight issue in that the code for setting up a giveaway
blithely assumed that there is always at least one entry in the
giveaway table, which there most certainly is not, particularly right
after you drop the table.

It is now smart enough to short circuit the tests it does if nothing was
found.