Set up an event bridge to make things less manky
As a result of a terrible circular dependency, this sets up a global
event bridge (term shamefully stolen from Axel669) that allows the
different systems in the project to listen for and raise events so that
they can communicate with each other.

This resolves not only the circular dependency but also makes the code
cleaner by having a clearer separation of concerns.