Flesh out the data and control flow for giveaways
This sets up some stub back end code and the necessary changes to the
front end code to be able to start, stop, pause and unpause a giveaway.

There's still work to be done here; countdowns are not happening, the
elapsed time is not updated, etc. However this ties all of the
appropriate bits and bobs together.