Use the event bridge to notify of twitch auths
This uses the event bridge added recently and gives support for it to
the Twitch authorization code, which can now emit events that tell the
other systems if a Twitch authorization has been added or dropped,
allowing them to take any actions they might need.

This affects the Twitch Chat and PubSub modules, which can now set up or
tear down without the Twitch code having to interface with them
directly.