Make cancellation confirmations more explicit
Previously the cancel button would directly cancel a running giveaway.

Since that is hyper dangerous and prone to make people extremely unhappy
at some point, we change that so that now when you click the button it
will expose a confirmation button that must also be clicked. After a
timeout, the buttons will revert if the confirmation is not clicked.