Include a userId in the giveaways table
The thought occurs that since the user can connect and disconnect their
account at any time, it's possible that they might swap accounts while
they're doing it.

Currently everything is predicated on a single user doing everything,
but what if they swap? what happens to the existing giveaways?

In order to forestall any issues, giveaways are now marked with the
userID of the person that created them so that we can be sure that only
giveaways that match the current user are considered "valid".