Handle twitch auth/deauth in the giveaway handlers
This change fleshes out the previously made changes to the giveaway
code so that we can listen for twitch auth and death events and either
start or stop a giveaway in response.

This uses Twitch auth so that we can make sure that the records we're
using match the currently authenticated user.