Implement starting a giveaway from the UI
This implements all of the code needed to have the user facing panel
prompt you for a duration and then kick off a new giveaway using it.

This leverages the existing code that finds in-progress giveaways at
startup time by creating a new database record and then invoking that
method to do the actual work of the giveaway starting.