Implement pausing and resuming of giveaways
This fleshes out the back end code that's required to pause the timer of
an ongoing giveaway and un-pause it again when desired, including after
startup if a giveaway was running.

This change only needs to be on the back end because all of the
infrastructure that we need in the front end is already done.