Include the ability to cancel a giveaway
This finishes the initial implementation of our panel controls by making
sure that we can cancel an ongoing giveaway if desired.

This is implemented in a way that marks the giveaway as cancelled, so
while the user interface says that it's not undoable, it technically
actually is, with a bit of manual work.