Refactor the client side layout
In order to make things a little bit cleaner and nicer to look at and
work with, this is a simple refactor of the code files for the front
end, including the panels and the overlay, to make the layout better.

This includes renaming the CSS files in question and changing how they
are brought in so that parcel will watch them properly, as well as
splitting the script out for everything to make things look a little
nicer as well.