Implement tracking of bits and subs
This is the initial implementation of the code for tracking when a user
gifts a sub or throws bits while a giveaway is in action.

We implement a cached version of the database contents in which we store
the list of people that have contributed to this particular giveaway
and how they have contributed.

Incoming messages for bits and subs will update both the local cache and
the database at once.