Include a helper function for making object ID's
We're transitioning the database to a system where we use textual object
ID's instead of the numeric rowID to get around some failings in
Trilogy and it's lack of ability to make a column autoincrement without
it also being the primary key.

This implements a simple helper function for creating object ID's.