Swap our tables to use the new objectID
This changes all code that creates, queries and removes entries from any
database table to use our new object ID instead of the rowID as the
key.

This allows the gifters table to have a composite primary key that
covers the giveaway id and the giveaway user.

As a part of this, we now support schema options for all tables,
although some are blank.