Fix a horrible logic issue in the handling code
There was a horrible logic bomb in the code that updates the database
in which if multiple requests arrive at the same time, they may not
sequence correctly due to the interaction between threads and the event
loop.

The solution is to be more smarter and update the database only after
all the tracking information in the cache is up to date, so that the
database can sequence the writes but the logic has up to date
information to make correct decisions.