Make all database writes distinct
Our code currently dumps writes to the database as they come in, and in
the case that there are several back to back updates for the same
giveaway user, we may end up queuing multiple writes for the same
object because we always use the same reference.

This change includes a simple shallow clone of the object so that all
writes are distinct.