Fix the CSS location for the overlay
For unknown brain fart related reasons (probably) the CSS for the
overlay was put into the common area instead of in the overlay area.

Probably this is because I thought there might be some need for common
overlay CSS, but there's gonna be only one, so that is not actually
needed any longer.