Transmit gift information
This implements a function that can be used to transmit an overlay info
update to any connected client; the update will happen for either bits,
subs or both as determined by arguments.

This change also wraps in a few call points for this to occur at,
although more work is needed in this area.