Allow for single socket updates
In order to allow incoming connections to get updates on the current
leaderboard information right when they connect, this updates the
function that does the transmit so that it can send to a specific
socket instead of everyone.

When sending to a single socket, the update happens right away, instead
of being debounced as it would be if it was a general update.