Include a new config item to set gifter view size
We want to be able to control the number of people that appear on the
leaderboards for both bits and subs, and we don't want to hardcode
that because that would make us all very, very sad.

So, there is a configuration option for both (each can be tweaked as
desired), which will be sent up to any page that connects to use and
needs websockets for communication, so that they can take advantage of
it if desired.