Set up some simple initial animation examples
This sets us up for a couple of different types of animations.

Firstly, when the list initially populates, the people in it bounce into
position from the right hand side of the screen, setting up the initial
display.

Secondly, whenever an update happens, the first and last items in the
list swap positions, just to show that effect and how to achieve it
using the gsap Flip plugin.