Alter the style of the JS for better positioning
In order to handle animations better going forward, we're adding in a
class that will absolutely position things relative to the parent, and
yes that will probably always be confusing to me.

In order to make the leader boxes look correct, this also requires a
margin change so that everything lines up, since the only way to make
the divs for the gift boxes not wrap is to set the left to 0
apparently.