Make the leaderboard headers dynamic
The style changes we made to the leaderboard items stops the header
from growing with the width of the child items automatically.

This change includes a simple helper that can size the header to the
width of the widest child in order to return the visual style that we're
interested in.