Update prisma and use a new font loader
In order to solve our weird layout problems that were getting in the way
of our animations, this brings in a new WebFont loader so that we can
defer the setup of the application until the fonts have been fully
loaded, since the font used changes the size of things.

Along the way we also upgraded Prisma to the latest version.