Finalize the initial animation for the overlay
This change pulls in our test logic and applies it to the actual update
data and stored leaderboard data (if any) so that we actually animate
people entering and leaving the list, list items changing positions,
and so on.

This is the initial version; more factoring is needed for this code,
and we want to clean up the animations somewhat as well.