Add a role system to our sockets
This change alters our contract with the client side code so that when
it connects, it must send us a message that tells us its "role" before
we will acknowledge that it exists.

We track all sockets, plus sockets by role, allowing a future update to
let you send a message to a specific subset of clients.