Include the drop game in the overlay
By popular request, this inlines the drop game from OdatNurd's channel
into the overlay. This is also a demonstration on how you could connect
the main drop game into your own back end infrastructure using
websocket communications (which is what the version in obotnurd does
anyway, but using a framework).

This is currently mildly hacky because it requires the function for
saying things in chat to be passed to the drop game setup in order to
avoid a nasty circular dependency issue.