Provide for a background on the leaderboards
This is something that has been pending for a while but there did not
seem to be a particularly easy way to get it set up. Overall what we
want is for the leaderboards to have some sort of background on them
that will make it easier to see them.

The child elements are positioned absolutely and so they are outside of
the flow, but since we know the exact maximum size of either board by
knowing the size of the elements and the total that there can be at
most, it's an easy matter to force the size as appropriate.