Make the announcement texts configurable
This enhances the existing code that announces when giveaway status is
changing so that the text that's announced is user configurable instead
of being hard coded.

In addition the announcement call has been modified internally to use
an action, while we try to figure out why the announce command is not
working the way that it's supposed to here.