Update the CSS to include a background on the time
This was missed in the previous update, but we want to make sure that
the giveaway countdown clock is also visible, so it should have the
possibility of a background on it.

For now this is just a placeholder, like the others, and can be styled
to look much better than it currently does.