Refactor the back end data model we're using
This changes how the back end code handles the giveaway data we store
and transmits it away.

The major part of this is that we now always have a "current" giveaway
even if it's one that has previously elapsed or was cancelled. The idea
is that when a giveaway ends, we should keep representing that data
until the next one occurs, so that the person running the giveaway can
see the results.

As a part of this, various things in the code were cleaned up, such as
changing the names of variables so that they better embody what the code
is now doing.