Include some common client side utility functions
All of our upcoming changes are predicated on the ability of the
overlays and other related panels to know the state of a giveaway,
such as if it's running or it's remaining duration.

This sets up some common code to be used by all of them so we don't
have to keep reinventing the wheel.