Include a results panel
For the purposes of monitoring the giveaway while it's in action without
having to try and squint at the leaderboard in the overlay, this adds
in a new results panel that will allow seeing the full list of
participants in any given giveaway leaderboard.

Unlike the overlay, this will display all of the participants and not
just the top however many are configured.