Update the debounce time on leader updates
Previously our leaderboard debounce time was 1 seconds between updates,
but it can take as much as 5 seconds for the animations in the overlay
to play out (as a worst case scenario).

We were seeing some issues in the console related to items not being
able to be removed from the DOM when animations finish, which I think
is a consequence of the updates not being throttled correctly.

This ends up being an issue in cases where we're manually and really
hammering on the button, forcing updates faster than the front end can
probably handle them.