Include controls that allow for adjusting data
Add in controls in the giveaway panel so that it's possible to adjust
the duration of a giveaway while one is running (to add or remove time
from the overall duration).

In addition there are also controls that allow you to add some number of
bits and subs to any twitch user that you like.

At the present time this is intended to only *add* and not *remove*,
though this could be done as well with more back end work.